Change file structure, add comments for inhabitedness.rs
